Update DOC.md
Questions:
Am I accessing most of this stuff from the command line? Or am I altering a file? If so, what file?

What kind of functionalities does "main" have, and why would I want to change it?

Do I understand the relationship between /sim and /anlys? Here's my attempt to explain: It seems that sim is managing the "under the hood" work, through delegation of commands to other parts of the program. And that /anlys is allowing the program to "observe" what is going on, so that /io can output the information.